move Label::ScreenTransform to own header
